requirements.txt
ttkbootstrap
pyinstaller
pillow
keyboard